prep for table
We do the same thing as with the name cards, were we first display empty entries, as if we try to load these right away, there is noticable stutter in the animation.
